// src/app/shared/models/lab.ts
export interface Lab {
    id?: number;
    patient_name: string ;
    amount_blood:string;
    hiv: string;
    pressure: string;
    weight:string;
    // id!: number;
    // patient_name!: String ;
    // amount_blood!: String ;
    // hiv!:String;
    // pressure!: String;
    // weight!: String ;
   
}